handle NIL namespace delimiters, update docs
We experience the following NAMESPACE command/response on a yahoo.co.jp
test account:

```
W6 NAMESPACE
* NAMESPACE (("" NIL)) NIL NIL
W6 OK NAMESPACE completed
```

The server greeting is:
```
* OK [CAPABILITY IMAP4rev1 ID NAMESPACE UIDPLUS LITERAL+ CHILDREN XAPPLEPUSHSERVICE AUTH=PLAIN AUTH=LOGIN] IMAP4rev1 imapgate-0.7.68_11_1.61475 imap515.mail.kks.yahoo.co.jp
```

The grammar in https://tools.ietf.org/html/rfc2342 says this is legal,
although I'm not seeing any examples or semantics:

   Namespace = nil / "(" 1*( "(" string SP  (<"> QUOTED_CHAR <"> /
      nil) *(Namespace_Response_Extension) ")" ) ")"

I've opted to have the value pass through as null rather than normalizing it.
This is inconsistent with browserbox's current behaviour for LIST/LSUB in this
case (which I submitted the patches for).  In those cases we normalize to '/',
which I didn't have a great reason for, but I also didn't know what else to do.
It may make sense to revisit that decision in the future.  I've updated the
docs to reflect the current behaviour and that it may change.

In this case it seems like if we're reporting the namespace at all we might as
well report a falsey value so any consuming logic can know that the value is
not to be trusted and should use the delimiter reported by LIST/LSUB.  I do
think we want to report the namespace since the root of "" is useful.

For reference, this is what LIST/LSUB looks like for the (freshly created)
account:

```
W4 LIST
* LIST (\\NoInferiors \\HasNoChildren) "/" "Bulk Mail"
* LIST (\\NoInferiors \\HasNoChildren) "/" "Draft"
* LIST (\\NoInferiors \\HasNoChildren) "/" "Inbox"
* LIST (\\NoInferiors \\HasNoChildren) "/" "Sent"
* LIST (\\NoInferiors \\HasNoChildren) "/" "Trash"
W4 OK LIST completed
W5 LSUB
* LSUB (\\NoInferiors \\HasNoChildren) "/" "Bulk Mail"
* LSUB (\\NoInferiors \\HasNoChildren) "/" "Draft"
* LSUB (\\NoInferiors \\HasNoChildren) "/" "Inbox"
* LSUB (\\NoInferiors \\HasNoChildren) "/" "Sent"
* LSUB (\\NoInferiors \\HasNoChildren) "/" "Trash"
W5 OK LSUB completed
```