Fixed a concurrency issue with precheck commands
Fixed a concurrency issue WRT prechecks with IMAP commands in them.
The issue is that the precheck mechanism depends on synchronous flow and
breakIdle was breaking that flow when multiple functions with commands
in their prechecks were called synchronously (instead of waiting for the
previous function's promise to be resolved).
Also added a test using setFlags(), which has selectMailbox() in a
precheck, as an example.