Fix lost callbacks on untagged responses
Commands (and their callbacks) are getting lost when the client receives
an untagged response from the server.
The problem is that when the client receives an untagged response it
calls _sendRequest which de-queues and overwrites the _currentCommand.
Therefore, when the actual response for the overwritten command arrives
it's command is lost and no callback is made and the promise left
hanging.
The solution seems to be to not let untagged responses interfere with
_currentCommand which is in-line with the IMAP IETF standard, see
section 2.2.1-2.