Add timeout to incoming data. (#139)
Data is received in (max) 4KB packets in _onData. _onData is currently clearing the error timeout set in send which means that if a connection breaks after the first packet of a response the client will hang.

These changes reset the timeout on every packet and only clear it when an entire command has arrived and we're sure the command's callback will be called.