Fixed attribute length condition in listMailBoxes
In ef1533662ad15f1dd5eb674f29fe0b5969533f1f a condition for the
number of attributes of the untagged response for the
listMailboxes command was changed from 'attr.length < 3' to
'!attr.length < 3', which is equivalent to 'attr.length >= 3'.
If this condition is true, the code will skip the inbox. The RFC
specifies the response from LIST to be of the form:
`S: * LIST (\Noselect) "/" ~/Mail/foo`
Which has 3 parameters, so the condition will always evaluate to
true and skip the mailbox. The same applies to LSUB.

SEE: https://tools.ietf.org/html/rfc3501#section-7.2.2